Move array masking util to test_common folder
